refactor(tests): 测试目录结构性熵减 — 提取共享模式 + 正交拆分 consolidation 模块
1. conftest.py 新增 FakeMcpClient 工厂、FakeTextItem/FakeImageContent 辅助类
   及 patch_extraction_deps 辅助函数，消除 test_extraction_provider.py 中
   16 处内联 FakeClient 重复定义和 50 处 monkeypatch 重复调用；
2. test_extraction_provider.py 1,403→1,248 行（净减 155 行，-11%）；
3. test_consolidation_pipeline.py 1,369 行单体文件按 10 个 Test 类正交分解为
   consolidation/ 子目录下 9 个独立文件（conftest + 8 个测试文件），
   每个文件职责单一、边界清晰；
4. 全量回归 1,661 测试通过，无破坏性变更。
